+++ doc/design.md
@@ -13,24 +13,5 @@ Rec
 - Bytebuffers (Direct, Mapped)
 
 ## Interfaces:
-```typescript
-interface Record {}
 
-interface Source {
-    stream(): Stream<Record>
-    to(target: Target)
-    tee(tee: Tee): Source
-}
-
-interface Tee {
-    emit(record: Record)
-    source(): Source
-    to(target: Target) = source().to(target)
-}
-
-interface Target {
-    put(record: Record);
-    tee(tee: Tee): Target
-    putAll(source: Source) = source.stream.forEach(this::put)
-}
-```
+see `rec-v2/src/main/resources/Rec-v2.d.ts`
