--- rec-accessor/build.gradle
@@ -1,12 +1,0 @@
-
-apply plugin: 'java'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    compile project(":common")
-    testCompile 'junit:junit:4.11'
-}
-
