--- rec-accessor/src/main/java/net/kimleo/rec/accessor/lexer/FieldType.java
@@ -1,4 +1,0 @@
-package net.kimleo.rec.accessor.lexer;
-
-public interface FieldType {
-}
