+++ rec-sepval/src/main/java/net/kimleo/rec/sepval/parser/ParseConfig.java
@@ -4,8 +4,8 @@ public class ParseConfig {
     public final char delimiter;
     public final char escape;
 
-    public static final ParseConfig DEFAULT = new ParseConfig(',', '"');
-    public static final ParseConfig BY_PIPE = new ParseConfig('|', '`');
+    public static final ParseConfig DEFAULT = new ParseConfig();
+    public static final ParseConfig BY_PIPE = new ParseConfig('|');
 
     public ParseConfig(char delimiter, char escape) {
         this.delimiter = delimiter;
