+++ rec-sepval/src/main/java/net/kimleo/rec/sepval/parser/SimpleParser.java
@@ -12,7 +12,7 @@ public class SimpleParser {
     }
 
     public SimpleParser() {
-        this(new ParseConfig());
+        this(ParseConfig.DEFAULT);
     }
 
     public SepValEntry parse(String input) {
