+++ rec-v2/build.gradle
@@ -18,7 +18,6 @@ task fatJar(type: Jar) {
 dependencies {
     compile project(":common")
     compile project(":rec-sepval")
-    compile project(":rec-accessor")
     compile "org.mozilla:rhino:1.7.7.1"
     testCompile 'junit:junit:4.11'
 }
