+++ rec-v2/src/main/java/net/kimleo/rec/v2/accessor/lexer/FieldName.java
@@ -1,9 +1,9 @@
-package net.kimleo.rec.accessor.lexer;
+package net.kimleo.rec.v2.accessor.lexer;
 
-public class Placeholder implements FieldType {
+public class FieldName implements FieldType {
     private final String name;
 
-    public Placeholder(String name) {
+    FieldName(String name) {
         this.name = name;
     }
 
@@ -16,9 +16,9 @@ public class Placeholder implements FieldType {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
 
-        Placeholder that = (Placeholder) o;
+        FieldName fieldName = (FieldName) o;
 
-        return name != null ? name.equals(that.name) : that.name == null;
+        return name != null ? name.equals(fieldName.name) : fieldName.name == null;
     }
 
     @Override
