+++ rec-v2/src/main/java/net/kimleo/rec/v2/accessor/lexer/FieldType.java
@@ -1,0 +1,4 @@
+package net.kimleo.rec.v2.accessor.lexer;
+
+public interface FieldType {
+}
